# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopNow Electronics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
  /recommendations/homepage:
    get:
      operationId: getHomepageRecommendations
      tags: [recommendations]
      summary: All three recommendation widget types for homepage
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recommendation widgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageRecommendations"
  /recommendations/pdp/{productId}:
    get:
      operationId: getPdpRecommendations
      tags: [recommendations]
      summary: PDP recommendation widgets (frequently bought together, content-based, hybrid)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDP recommendation widgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdpRecommendations"
...
components:
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - brand
        - price
        - category
        - imageUrl
        - rating
        - reviewCount
        - inStock
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountPct:
          type: ["integer", "null"]
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        stockCount:
          type: ["integer", "null"]
        specs:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isDeal:
          type: boolean
    RecommendationWidget:
      type: object
      required:
        - type
        - title
        - subtitle
        - products
      properties:
        type:
          type: string
          enum: [content_based, collaborative, hybrid]
        title:
          type: string
        subtitle:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/RecommendedProduct"
    RecommendedProduct:
      type: object
      required:
        - product
        - reason
      properties:
        product:
          $ref: "#/components/schemas/Product"
        reason:
          type: string
    HomepageRecommendations:
      type: object
      required:
        - contentBased
        - collaborative
        - hybrid
      properties:
        contentBased:
          $ref: "#/components/schemas/RecommendationWidget"
        collaborative:
          $ref: "#/components/schemas/RecommendationWidget"
        hybrid:
          $ref: "#/components/schemas/RecommendationWidget"
    PdpRecommendations:
      type: object
      required:
        - frequentlyBoughtTogether
        - contentBased
        - hybrid
      properties:
        frequentlyBoughtTogether:
          $ref: "#/components/schemas/RecommendationWidget"
        contentBased:
          $ref: "#/components/schemas/RecommendationWidget"
        hybrid:
          $ref: "#/components/schemas/RecommendationWidget"
...